src/func/pyproject.toml
[project]
name = "diary-rag"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = "~=3.11.0"
readme = "README.md"
dependencies = [
    "azure-functions>=1.21.3,<2",
    "azure-cosmos>=4.7.0,<5",
    "azure-identity>=1.16.0,<2",
    "python-dotenv>=1.0.0,<2",
    "openai>=1.0.0,<2",
    "cryptography>=46.0.3,<47",
    "google-auth>=2.35.0,<3",
    "google-auth-oauthlib>=1.2.1,<2",
    "google-auth-httplib2>=0.2.0,<0.3",
    "google-api-python-client>=2.149.0,<3",
    "langchain-core>=1.0,<2",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.11.8",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 127
